fix router and move graph related parts to GraphView.vue
