+++ src/components/dialog/UnloadWindowConfirmDialog.vue
@@ -1,16 +1,18 @@
 <template>
-  <div>
-    <!-- This component does not render anything visible. It is used to confirm
-    the user wants to close the window, and if they do, it will call the
-    beforeunload event. -->
-  </div>
+  <!--
+  UnloadWindowConfirmDialog: This component does not render
+  anything visible. It is used to confirm the user wants to
+  close the window, and if they do, it will call the
+  beforeunload event.
+  -->
 </template>
 
 <script setup lang="ts">
 import { useSettingStore } from '@/stores/settingStore'
-import { onMounted, onUnmounted } from 'vue'
+import { onMounted, onBeforeUnmount } from 'vue'
 
 const settingStore = useSettingStore()
+
 const handleBeforeUnload = (event: BeforeUnloadEvent) => {
   if (settingStore.get('Comfy.Window.UnloadConfirmation')) {
     event.preventDefault()
@@ -23,7 +25,7 @@ onMounted(() => {
   window.addEventListener('beforeunload', handleBeforeUnload)
 })
 
-onUnmounted(() => {
+onBeforeUnmount(() => {
   window.removeEventListener('beforeunload', handleBeforeUnload)
 })
 </script>
